make editors javascript work (admin)
