+++ app/packs/src/decidim/decidim_awesome/awesome_application.js
@@ -1,3 +1,3 @@
 import "src/decidim/decidim_awesome/proposals/images"
-import "src/decidim/decidim_awesome/editors/quill_editor"
+import "src/decidim/decidim_awesome/editors/editor"
 import "src/decidim/decidim_awesome/forms/autosave"
