+++ app/packs/stylesheets/decidim/decidim_awesome/awesome_admin.scss
@@ -2,8 +2,9 @@
 @import "stylesheets/decidim/decidim_awesome/admin/constraints";
 @import "stylesheets/decidim/decidim_awesome/admin/custom_fields";
 @import "stylesheets/decidim/decidim_awesome/admin/auto_edits";
-// @import "stylesheets/decidim/decidim_awesome/editors/quill_editor";
 @import "stylesheets/decidim/decidim_awesome/editors/markdown_editor";
+@import "stylesheets/decidim/decidim_awesome/editors/quill_editor";
+
 
 .platoniq-credits{
   position: absolute;
