+++ app/packs/stylesheets/decidim/decidim_awesome/editors/markdown_editor.scss
@@ -1,3 +1,5 @@
+@import "inscrybmde/src/css/inscrybmde.css";
+
 // editor tweaks
 //
 .editor .CodeMirror{
